refactor PostSingle files into separate folder
