+++ src/post/postSingle/PostSingleModal.scss
